Create README.md
I'm trying to flash my digi-uno to the current softward, but not sure which one to use.  and the Wled_0.15.0_ESP32.bin.  It flashed fine, but now it won't connect to the internet.  I bought the digi-uno a couple years ago. Does that mean it can't handel the updated softward?